FLOW-482 Now only deletes the output file if it exists (re: code review).

+++ uploadConfigPanel.sh
@@ -36,4 +36,4 @@ else
   cat uploadConfigPanelResponse.txt
 fi
 
-rm uploadConfigPanelResponse.txt
+[ -e uploadConfigPanelResponse.txt ] && rm uploadConfigPanelResponse.txt
